backend/pyproject.toml
[project]
name = "bumblebee-backend"
version = "0.1.0"
description = "Bumblebee IDE backend - Visual Logic Engine"
requires-python = ">=3.12"
dependencies = [
    "falkordb>=1.0.0",
    "tree-sitter>=0.23.0",
    "tree-sitter-python>=0.23.0",
    "watchdog>=4.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.0",
    "websockets>=12.0",
    "httpx>=0.27",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "black>=24.0",
    "isort>=5.13",
    "pylint>=3.2",
    "mypy>=1.10",
    "httpx>=0.27",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[tool.pylint.format]
max-line-length = 120

[tool.pylint."messages control"]
disable = ["missing-module-docstring"]

[tool.mypy]
strict = true
python_version = "3.12"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
